# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
  handlers.RefreshTokenRequestBody:
    properties:
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    required:
    - refreshToken
    type: object
  handlers.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
  handlers.TokensResponseBody:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
...
host: gw.verifire.com/auth
info:
  contact:
    name: Artem Kostenko
    url: https://github.com/aerosystems
  description: A mandatory part of any microservice infrastructure of a modern WEB
    application, which is responsible for user authentication and authorization.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Auth Service
  version: 1.0.7
paths:
...
  /v1/token/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: raw request body, should contain Refresh Token
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.TokensResponseBody'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: refresh a pair of JWT tokens
      tags:
      - auth
  /v1/token/validate:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: validate token
      tags:
      - api-gateway-special
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Should contain Access JWT Token, with the Bearer started
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"